+++ CHANGES.md
@@ -4,6 +4,7 @@ Change Log
 #### next release (8.1.14)
 * Reimplement option to zoom on item when adding it to workbench, `zoomOnAddToWorkbench` is added to `MappableTraits`.
 * Update terria-js cesium to `1.81.3` 
+* Fix `GeoJsonMixin.legend` for Cesium primitives (`TableMixin` will now check for `super.legends`, and it will not return legends unless it has data)
 * [The next improvement]
 
 #### 8.1.13